Set org name from org domain
