+++ packages/server/customer-os-common-module/utils/core_utils.go
@@ -3,7 +3,6 @@ package utils
 import (
 	"encoding/json"
 	"math/rand"
-	"net/url"
 	"reflect"
 	"runtime"
 	"strings"
@@ -12,7 +11,6 @@ import (
 	"github.com/neo4j/neo4j-go-driver/v5/neo4j/dbtype"
 	"github.com/pkg/errors"
 	"github.com/sirupsen/logrus"
-	"golang.org/x/net/publicsuffix"
 )
 
 type Pair[T, U any] struct {
@@ -333,66 +331,6 @@ func ToJson(obj any) (string, error) {
 		return "", err
 	}
 	return string(outputJson), nil
-}
-
-func ExtractDomainFromEmail(email string) string {
-	emailParts := strings.Split(email, "@")
-	if len(emailParts) != 2 {
-		return ""
-	}
-	return ExtractDomain(emailParts[1])
-}
-
-func ExtractDomain(input string) string {
-	if !strings.Contains(input, ".") {
-		return ""
-	}
-
-	hostname := extractHostname(strings.TrimSpace(strings.ToLower(input)))
-
-	domain, err := publicsuffix.EffectiveTLDPlusOne(hostname)
-	if err != nil {
-		return ""
-	}
-
-	if IsValidTLD(domain) {
-		return domain
-	}
-	return ""
-}
-
-func extractHostname(inputURL string) string {
-	// Prepend "http://" if the URL doesn't start with a scheme
-	if !strings.HasPrefix(inputURL, "http://") && !strings.HasPrefix(inputURL, "https://") {
-		inputURL = "http://" + inputURL
-	}
-
-	// Parse the URL
-	u, err := url.Parse(inputURL)
-	if err != nil {
-		return ""
-	}
-
-	// Extract and return the hostname (domain)
-	hostname := u.Hostname()
-
-	// Remove "www." if it exists
-	if strings.HasPrefix(hostname, "www.") {
-		hostname = hostname[4:] // Remove the first 4 characters ("www.")
-	}
-
-	return strings.ToLower(hostname)
-}
-
-func IsValidTLD(input string) bool {
-	etld, im := publicsuffix.PublicSuffix(input)
-	var validtld = false
-	if im { // ICANN managed
-		validtld = true
-	} else if strings.IndexByte(etld, '.') >= 0 { // privately managed
-		validtld = true
-	}
-	return validtld
 }
 
 func IsEmptyString(s *string) bool {
